Post-push fix for MDEV-11556: Make the debug variable UINT.
Sometimes innodb_data_file_size_debug was reported as INT UNSIGNED
instead of BIGINT UNSIGNED. Make it uint instead of ulong to get
a more deterministic result.

+++ storage/xtradb/srv/srv0start.cc
@@ -132,7 +132,7 @@ UNIV_INTERN ibool	srv_was_started = FALSE;
 static ibool		srv_start_has_been_called = FALSE;
 #ifdef UNIV_DEBUG
 /** InnoDB system tablespace to set during recovery */
-UNIV_INTERN ulong	srv_sys_space_size_debug;
+UNIV_INTERN uint	srv_sys_space_size_debug;
 #endif /* UNIV_DEBUG */
 
 /** At a shutdown this value climbs from SRV_SHUTDOWN_NONE to
